# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sovereign Office of the Chief Justice & Trustee API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trust/instruments:
    get:
      operationId: listInstruments
      tags: [instruments]
      summary: List all trust instruments
      responses:
        "200":
          description: List of instruments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrustInstrumentSummary"
    post:
      operationId: createInstrument
      tags: [instruments]
      summary: Create and generate a recorder-compliant trust instrument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrustInstrumentInput"
      responses:
        "201":
          description: Created instrument with PDF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrustInstrumentResponse"
...
  /trust/instruments/{id}:
    get:
      operationId: getInstrument
      tags: [instruments]
      summary: Get trust instrument by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trust instrument
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrustInstrumentDetail"
        "404":
          description: Not found
  /trust/instruments/{id}/file:
    post:
      operationId: fileInstrument
      tags: [instruments]
      summary: File an instrument with a county recorder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilingInput"
      responses:
        "201":
          description: Filing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrustFiling"
  /trust/instruments/{id}/submit:
    post:
      operationId: submitInstrument
      tags: [instruments]
      summary: Submit instrument to recorder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrustFiling"
...
components:
  schemas:
...
    TrustInstrumentSummary:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        instrumentType: { type: string }
        status: { type: string }
        jurisdiction: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        county: { type: ["string", "null"] }
        landClassification: { type: ["string", "null"] }
        pdfUrl: { type: ["string", "null"] }
        validationErrors: { type: ["array", "null"], items: { type: string } }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, title, instrumentType, status, createdAt, updatedAt]
    TrustInstrumentDetail:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        instrumentType: { type: string }
        content: { type: string }
        status: { type: string }
        jurisdiction: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        county: { type: ["string", "null"] }
        landClassification: { type: ["string", "null"] }
        pdfUrl: { type: ["string", "null"] }
        validationErrors: { type: ["array", "null"], items: { type: string } }
        trusteeNotes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, title, instrumentType, content, status, createdAt, updatedAt]
    TrustInstrumentInput:
      type: object
      properties:
        type: { type: string, description: "Instrument type (e.g. trust_deed, allotment_lease)" }
        title: { type: string }
        parties: { type: ["array", "null"], items: { type: string } }
        landDescription: { type: string }
        jurisdiction: { type: string }
        state: { type: string }
        templateKey: { type: string, description: "Use a built-in template key" }
        templateVariables:
          type: array
          items:
            type: object
            properties:
              key: { type: string }
              value: { type: string }
        indianLandProtection: { type: boolean }
        trustStatus: { type: boolean }
        federalPreemption: { type: boolean }
        tribalJurisdiction: { type: boolean }
        trusteeNotes: { type: string }
        recorderMetadata:
          type: object
          properties:
            returnAddress: { type: string }
            apn: { type: string }
            county: { type: string }
            state: { type: string }
            documentType: { type: string }
            requiresNotary: { type: boolean }
    TrustInstrumentResponse:
      type: object
      properties:
        instrument:
          $ref: "#/components/schemas/TrustInstrumentDetail"
        validation:
          type: object
          properties:
            errors: { type: array, items: { type: string } }
            warnings: { type: array, items: { type: string } }
            valid: { type: boolean }
        pdf:
          type: object
          properties:
            pages: { type: integer }
            checksum: { type: string }
            generatedAt: { type: string }
            downloadUrl: { type: string }
    FilingInput:
      type: object
      properties:
        county: { type: string }
        state: { type: string }
        instrumentId: { type: integer }
        documentType: { type: string }
        notes: { type: string }
      required: [county, state]
    SubmitInput:
      type: object
      properties:
        county: { type: string }
        state: { type: string }
    TrustFiling:
      type: object
      properties:
        id: { type: integer }
        instrumentId: { type: integer }
        county: { type: string }
        state: { type: string }
        filingStatus: { type: string }
        filingNumber: { type: ["string", "null"] }
        documentType: { type: ["string", "null"] }
        trustStatus: { type: ["string", "null"] }
        landClassification: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        submittedAt: { type: ["string", "null"] }
        acceptedAt: { type: ["string", "null"] }
        rejectedAt: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, instrumentId, county, state, filingStatus, createdAt, updatedAt]
...