# lib/api-zod/src/generated/types/trustInstrumentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sovereign Office of the Chief Justice & Trustee API
 * OpenAPI spec version: 0.1.0
 */
import type { TrustInstrumentInputRecorderMetadata } from "./trustInstrumentInputRecorderMetadata";
import type { TrustInstrumentInputTemplateVariablesItem } from "./trustInstrumentInputTemplateVariablesItem";

export interface TrustInstrumentInput {
  /** Instrument type (e.g. trust_deed, allotment_lease) */
  type?: string;
  title?: string;
  /** @nullable */
  parties?: string[] | null;
  landDescription?: string;
  jurisdiction?: string;
  state?: string;
  /** Use a built-in template key */
  templateKey?: string;
  templateVariables?: TrustInstrumentInputTemplateVariablesItem[];
  indianLandProtection?: boolean;
  trustStatus?: boolean;
  federalPreemption?: boolean;
  tribalJurisdiction?: boolean;
  trusteeNotes?: string;
  recorderMetadata?: TrustInstrumentInputRecorderMetadata;
}