# lib/api-zod/src/generated/types/trustInstrumentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sovereign Office of the Chief Justice & Trustee API
 * OpenAPI spec version: 0.1.0
 */
import type { TrustInstrumentDetail } from "./trustInstrumentDetail";
import type { TrustInstrumentResponsePdf } from "./trustInstrumentResponsePdf";
import type { TrustInstrumentResponseValidation } from "./trustInstrumentResponseValidation";

export interface TrustInstrumentResponse {
  instrument?: TrustInstrumentDetail;
  validation?: TrustInstrumentResponseValidation;
  pdf?: TrustInstrumentResponsePdf;
}